+++ c/nla-digbench/fermat2-ll.c
@@ -22,13 +22,13 @@ int main() {
     A = __VERIFIER_nondet_int();
     R = __VERIFIER_nondet_int();
     //assume_abort_if_not(A >= 1);
-    assume_abort_if_not((long long) (R - 1) * (R - 1) < A);
+    assume_abort_if_not(((long long) R - 1) * ((long long) R - 1) < A);
     //assume_abort_if_not(A <= R * R);
     assume_abort_if_not(A % 2 == 1);
 
-    u = (long long) 2 * R + 1;
+    u = ((long long) 2 * R) + 1;
     v = 1;
-    r = (long long) R * R - A;
+    r = ((long long) R * R) - A;
 
     while (1) {
         __VERIFIER_assert(4*(A+r) == u*u - v*v - 2*u + 2*v);
@@ -44,6 +44,6 @@ int main() {
     }
 
     //return  (u - v) / 2;
-    __VERIFIER_assert(4*A == u*u - v*v  - 2*u + 2*v);
+    __VERIFIER_assert(((long long) 4*A) == u*u - v*v  - 2*u + 2*v);
     return 0;
 }
